refactoring packages and using generics
