+++ spring-5-reactive/src/main/java/com/baeldung/reactive/filters/ExampleWebFilter.java
@@ -1,4 +1,4 @@
-package com.baeldung.reactive.filter;
+package com.baeldung.reactive.filters;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.server.ServerWebExchange;
