+++ spring-5-reactive/src/main/java/com/baeldung/reactive/filters/PlayerHandler.java
@@ -1,4 +1,4 @@
-package com.baeldung.reactive.handler;
+package com.baeldung.reactive.filters;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.server.ServerRequest;
@@ -8,9 +8,9 @@ import reactor.core.publisher.Mono;
 import static org.springframework.web.reactive.function.server.ServerResponse.ok;
 
 @Component
-public class PlayerHandler {
+class PlayerHandler {
 
-    public Mono<ServerResponse> getName(ServerRequest request) {
+    Mono<ServerResponse> getName(ServerRequest request) {
         Mono<String> name = Mono.just(request.pathVariable("name"));
         return ok().body(name, String.class);
     }
